python-main/facs/requirements.txt
anyio==3.6.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
fastapi==0.95.2
greenlet==2.0.1
h11==0.14.0
idna==3.4
importlib-metadata==5.0.0
mysql-connector-python==8.0.31
numpy==1.21.6
pandas==1.3.5
passlib==1.7.4
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
python-dateutil==2.8.2
pytest==7.2.0
pytest-mock==3.10.0
pytz==2022.6
requests==2.28.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.43
starlette==0.27.0
typing==3.7.4.3
typing-extensions==4.4.0
urllib3==1.26.13
uvicorn==0.19.0
zipp==3.10.0
httpx==0.24.0